simplify create into stages
